+++ srcs_bonus/game/textures_sprites.c
@@ -14,7 +14,7 @@
 
 void		frame_add_sprite(int frame_x, t_cub *cub)
 {
-	double		d_angle;
+	float		d_angle;
 	t_sprites	*spr;
 
 	d_angle = cub->cam->cam_angle_pitch / cub->win->y;
@@ -37,17 +37,17 @@ void		frame_add_sprite(int frame_x, t_cub *cub)
 	}
 }
 
-void		frame_add_sprite_one(int frame_x, double d_angle,
+void		frame_add_sprite_one(int frame_x, float d_angle,
 		t_sprites *spr, t_cub *cub)
 {
-	double			f_angl;
-	double			c_angl;
-	double			angle;
+	float			f_angl;
+	float			c_angl;
+	float			angle;
 	int				frame_y;
 	unsigned int	pixel;
 
-	f_angl = atan((-cub->cam->z) / spr->len_to_sp);
-	c_angl = atan((cub->map->blk_z - cub->cam->z) / spr->len_to_sp);
+	f_angl = atanf((-cub->cam->z) / spr->len_to_sp);
+	c_angl = atanf((cub->map->blk_z - cub->cam->z) / spr->len_to_sp);
 	angle = cub->cam->cam_direction_pitch + cub->cam->cam_angle_pitch / 2;
 	frame_y = -1;
 	while (spr->sp_x >= 0 && ++frame_y < cub->win->y)
@@ -55,7 +55,7 @@ void		frame_add_sprite_one(int frame_x, double d_angle,
 		if (angle < c_angl && angle > f_angl)
 		{
 			pixel = get_pixel_wall(spr->sp_x,
-								   1. - get_y_wall(angle, spr->len_to_sp, cub), cub,
+								   1.f - get_y_wall(angle, spr->len_to_sp, cub), cub,
 								   1);
 			if (pixel != 0)
 				((unsigned int*)cub->frm_0->data)[frame_y * cub->win->x +
